FIX : 백엔드 구조 변경 및 auth API를 명세에 맞게 변경
라우팅 함수가 점점 길어져 코드 관리가 힘들어지는 것을 막기 위해
라우팅 함수를 컨트롤러 폴더로 빼서 관리하도록 구조를 변경하였습니다.

추가적으로 로그인/회원가입시 반환 데이터를 API 문서대로 반환하도록
변수명을 user에서 data로 수정하였습니다.
