+++ WEB(BE)/controller/ctrl.common.js
@@ -1,0 +1,8 @@
+"use strict";
+
+export const Result = function() {
+    this.success = false;
+    this.data = null;
+    this.err_code = null;
+    this.err_msg = null;
+};
